month-1/week-1/day-2/pyproject.toml
[project]
name = "day-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
